+++ lib/runners/processor/metrics_codeclone.rb
@@ -1,0 +1,4 @@
+module Runners
+  class Processor::MetricsCodeClone < PmdCpd
+  end
+end