+++ test/config_test.rb
@@ -76,6 +76,7 @@ class ConfigTest < Minitest::Test
           jshint: nil,
           ktlint: nil,
           languagetool: nil,
+          metrics_codeclone: nil,
           misspell: nil,
           phinder: nil,
           phpmd: nil,