+++ CHANGELOG.md
@@ -12,6 +12,7 @@ All notable changes to this project will be documented in this file.
 - **RuboCop** Add supported gems [#1834](https://github.com/sider/runners/pull/1834)
 - Bump devon_rex images from 2.28.0 to 2.30.0 [#1859](https://github.com/sider/runners/pull/1859) [#1864](https://github.com/sider/runners/pull/1864) [#1875](https://github.com/sider/runners/pull/1875)
 - Gemification [#1858](https://github.com/sider/runners/pull/1858)
+- **Metrics Code Clone** [#TBD](https://github.com/sider/runners/pull/TBD)
 
 ## 0.39.3